+++ internal/api/web/_app/immutable/entry/start.REAUJ276.js
@@ -1,0 +1,1 @@
+import{o as e,t}from"../chunks/C7DqvD_f.js";export{e as load_css,t as start};